refactor: decouple connection status from useMCPServerManager with useMCPConnectionStatus
