+++ client/src/components/Chat/Input/MCPSubMenu.tsx
@@ -3,8 +3,8 @@ import * as Ariakit from '@ariakit/react';
 import { ChevronRight } from 'lucide-react';
 import { PinIcon, MCPIcon } from '@librechat/client';
 import MCPServerStatusIcon from '~/components/MCP/MCPServerStatusIcon';
-import { useMCPServerManager } from '~/hooks/MCP/useMCPServerManager';
 import MCPConfigDialog from '~/components/MCP/MCPConfigDialog';
+import { useMCPServerManager } from '~/hooks';
 import { cn } from '~/utils';
 
 interface MCPSubMenuProps {
