+++ client/src/hooks/MCP/index.ts
@@ -1,4 +1,5 @@
-export * from './useMCPSelect';
 export * from './useGetMCPTools';
+export * from './useMCPConnectionStatus';
+export * from './useMCPSelect';
 export * from './useVisibleTools';
 export { useMCPServerManager } from './useMCPServerManager';
