+++ client/src/hooks/MCP/useMCPServerManager.ts
@@ -9,8 +9,7 @@ import {
 } from 'librechat-data-provider/react-query';
 import type { TUpdateUserPlugins, TPlugin } from 'librechat-data-provider';
 import type { ConfigFieldDetail } from '~/common';
-import { useMCPConnectionStatusQuery } from '~/data-provider/Tools/queries';
-import { useLocalize, useMCPSelect, useGetMCPTools } from '~/hooks';
+import { useLocalize, useMCPSelect, useGetMCPTools, useMCPConnectionStatus } from '~/hooks';
 import { useGetStartupConfig } from '~/data-provider';
 
 interface ServerState {
@@ -83,13 +82,9 @@ export function useMCPServerManager({ conversationId }: { conversationId?: strin
     return initialStates;
   });
 
-  const { data: connectionStatusData } = useMCPConnectionStatusQuery({
+  const { connectionStatus } = useMCPConnectionStatus({
     enabled: !!startupConfig?.mcpServers && Object.keys(startupConfig.mcpServers).length > 0,
   });
-  const connectionStatus = useMemo(
-    () => connectionStatusData?.connectionStatus || {},
-    [connectionStatusData?.connectionStatus],
-  );
 
   /** Filter disconnected servers when values change, but only after initial load
    This prevents clearing selections on page refresh when servers haven't connected yet
@@ -97,7 +92,7 @@ export function useMCPServerManager({ conversationId }: { conversationId?: strin
   const hasInitialLoadCompleted = useRef(false);
 
   useEffect(() => {
-    if (!connectionStatusData || Object.keys(connectionStatus).length === 0) {
+    if (!connectionStatus || Object.keys(connectionStatus).length === 0) {
       return;
     }
 
@@ -115,7 +110,7 @@ export function useMCPServerManager({ conversationId }: { conversationId?: strin
     if (connectedSelected.length !== mcpValues.length) {
       setMCPValues(connectedSelected);
     }
-  }, [connectionStatus, connectionStatusData, mcpValues, setMCPValues]);
+  }, [connectionStatus, mcpValues, setMCPValues]);
 
   const updateServerState = useCallback((serverName: string, updates: Partial<ServerState>) => {
     setServerStates((prev) => {
@@ -351,7 +346,7 @@ export function useMCPServerManager({ conversationId }: { conversationId?: strin
           return;
         }
 
-        const serverStatus = connectionStatus[serverName];
+        const serverStatus = connectionStatus?.[serverName];
         if (serverStatus?.connectionState === 'connected') {
           connectedServers.push(serverName);
         } else {
@@ -381,7 +376,7 @@ export function useMCPServerManager({ conversationId }: { conversationId?: strin
         const filteredValues = currentValues.filter((name) => name !== serverName);
         setMCPValues(filteredValues);
       } else {
-        const serverStatus = connectionStatus[serverName];
+        const serverStatus = connectionStatus?.[serverName];
         if (serverStatus?.connectionState === 'connected') {
           setMCPValues([...currentValues, serverName]);
         } else {
@@ -455,7 +450,7 @@ export function useMCPServerManager({ conversationId }: { conversationId?: strin
   const getServerStatusIconProps = useCallback(
     (serverName: string) => {
       const tool = mcpToolDetails?.find((t) => t.name === serverName);
-      const serverStatus = connectionStatus[serverName];
+      const serverStatus = connectionStatus?.[serverName];
       const serverConfig = startupConfig?.mcpServers?.[serverName];
 
       const handleConfigClick = (e: React.MouseEvent) => {
@@ -532,7 +527,7 @@ export function useMCPServerManager({ conversationId }: { conversationId?: strin
 
     return {
       serverName: selectedToolForConfig.name,
-      serverStatus: connectionStatus[selectedToolForConfig.name],
+      serverStatus: connectionStatus?.[selectedToolForConfig.name],
       isOpen: isConfigModalOpen,
       onOpenChange: handleDialogOpenChange,
       fieldsSchema,
@@ -553,7 +548,6 @@ export function useMCPServerManager({ conversationId }: { conversationId?: strin
 
   return {
     configuredServers,
-    connectionStatus,
     initializeServer,
     cancelOAuthFlow,
     isInitializing,
